Updating ReadMe
This updated ReadMe considers the scope of the new project and includes an overview, instructions on running the code, the ideal output and analysis, and future considerations of the project. 
This file should be the final piece to the project, but feel free to add and update as necessary. 
Also, I was unable to name this README.md because that file already exists.